Splited telegram notifier into an api and a notifier, so it can be reused in the telegram publisher
